+++ hitsdb-20200615/src/Models/GetLindormInstanceResponseBody.php
@@ -6,6 +6,7 @@ namespace AlibabaCloud\SDK\Hitsdb\V20200615\Models;
 
 use AlibabaCloud\Dara\Model;
 use AlibabaCloud\SDK\Hitsdb\V20200615\Models\GetLindormInstanceResponseBody\engineList;
+use AlibabaCloud\SDK\Hitsdb\V20200615\Models\GetLindormInstanceResponseBody\singleZoneRiskAlert;
 
 class GetLindormInstanceResponseBody extends Model
 {
@@ -275,6 +276,11 @@ class GetLindormInstanceResponseBody extends Model
     public $serviceType;
 
     /**
+     * @var singleZoneRiskAlert
+     */
+    public $singleZoneRiskAlert;
+
+    /**
      * @var string
      */
     public $standbyVSwitchId;
@@ -352,6 +358,7 @@ class GetLindormInstanceResponseBody extends Model
         'requestId' => 'RequestId',
         'resourceGroupId' => 'ResourceGroupId',
         'serviceType' => 'ServiceType',
+        'singleZoneRiskAlert' => 'SingleZoneRiskAlert',
         'standbyVSwitchId' => 'StandbyVSwitchId',
         'standbyZoneId' => 'StandbyZoneId',
         'vpcId' => 'VpcId',
@@ -363,6 +370,9 @@ class GetLindormInstanceResponseBody extends Model
     {
         if (\is_array($this->engineList)) {
             Model::validateArray($this->engineList);
+        }
+        if (null !== $this->singleZoneRiskAlert) {
+            $this->singleZoneRiskAlert->validate();
         }
         parent::validate();
     }
@@ -587,6 +597,10 @@ class GetLindormInstanceResponseBody extends Model
 
         if (null !== $this->serviceType) {
             $res['ServiceType'] = $this->serviceType;
+        }
+
+        if (null !== $this->singleZoneRiskAlert) {
+            $res['SingleZoneRiskAlert'] = null !== $this->singleZoneRiskAlert ? $this->singleZoneRiskAlert->toArray($noStream) : $this->singleZoneRiskAlert;
         }
 
         if (null !== $this->standbyVSwitchId) {
@@ -837,6 +851,10 @@ class GetLindormInstanceResponseBody extends Model
 
         if (isset($map['ServiceType'])) {
             $model->serviceType = $map['ServiceType'];
+        }
+
+        if (isset($map['SingleZoneRiskAlert'])) {
+            $model->singleZoneRiskAlert = singleZoneRiskAlert::fromMap($map['SingleZoneRiskAlert']);
         }
 
         if (isset($map['StandbyVSwitchId'])) {